Added openmp. Changed long double to double (preparations for SSE)

+++ inc/Commons.h
@@ -2,6 +2,6 @@
 #define COMMONS_H
 #include <math.h>
 
-long double p2p_distance(long double a, long double b);
+double p2p_distance(double a, double b);
 
 #endif
