+++ inc/Mandelbrot.h
@@ -5,13 +5,13 @@
 
 void MandelbrotInitColorTable(void);
 
-void MandelbrotResetBounds(long double *lx,
-                           long double *rx,
-                           long double *ty,
-                           long double *by);
+void MandelbrotResetBounds(double *lx,
+                           double *rx,
+                           double *ty,
+                           double *by);
 
-uint32_t MandelbrotGetColor(long double x0,
-                            long double y0);
+uint32_t MandelbrotGetColor(double x0,
+                            double y0);
 
 
 #endif
